Moved metrics under pkg
